# src/main/resources/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Contract First API
  termsOfService: "#"
  contact:
    email: gabriel@alencar.site
  license:
    name: Xxxxx
    url: http://#
paths:
  v1/item/pagination:
    get:
      tags:
        - ItemController
      summary: Pagination
      operationId: pagination
      parameters:
        - name: order
          in: query
          description: order
          required: false
          schema:
            type: string
            default: DESC
        - name: orderby
          in: query
          description: field to order by
          required: false
          schema:
            type: string
            enum:
              - id
              - name
            default: name
        - name: page
          in: query
          description: number of page
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: limit
          in: query
          description: rows per page
          required: false
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationItemResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  v1/user/pagination:
    get:
      tags:
        - UserController
      summary: Pagination
      operationId: pagination
      parameters:
        - name: order
          in: query
          description: order
          required: false
          schema:
            type: string
            default: DESC
        - name: orderby
          in: query
          description: field to order by
          required: false
          schema:
            type: string
            enum:
              - id
              - name
            default: name
        - name: page
          in: query
          description: number of page
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: limit
          in: query
          description: rows per page
          required: false
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationUserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  v1/order/pagination:
    get:
      tags:
        - OrderController
      summary: Pagination
      operationId: pagination
      parameters:
        - name: order
          in: query
          description: order
          required: false
          schema:
            type: string
            default: DESC
        - name: orderby
          in: query
          description: field to order by
          required: false
          schema:
            type: string
            enum:
              - id
              - name
            default: name
        - name: page
          in: query
          description: number of page
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: limit
          in: query
          description: rows per page
          required: false
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationOrderResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  v1/stock-movement/pagination:
    get:
      tags:
        - StockMovementController
      summary: Pagination
      operationId: pagination
      parameters:
        - name: order
          in: query
          description: order
          required: false
          schema:
            type: string
            default: DESC
        - name: orderby
          in: query
          description: field to order by
          required: false
          schema:
            type: string
            enum:
              - id
              - name
            default: name
        - name: page
          in: query
          description: number of page
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: limit
          in: query
          description: rows per page
          required: false
          schema:
            type: integer
            format: int32
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginationStockMovementResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
servers:
  - url: https://localhost
  - url: http://localhost
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "401"
        message:
          type: string
          example: "Unauthorized"
...
    Item:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          example: 149467
          required: true
        name:
          type: string
          example: Xxxxxx
          required: true
...
    PaginationItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        totalPages:
          type: integer
          format: int32
          example: 50
        totalElements:
          type: integer
          format: int64
          example: 9266
        size:
          type: integer
          format: int32
          example: 100
        number:
          type: integer
          format: int32
          example: 0
...
    User:
      type: object
      required:
        - email
        - name
      properties:
        id:
          type: integer
          example: 149467
          required: true
        name:
          type: string
          example: Xxxxxx
          required: true
        email:
          type: string
          example: email@email.com
          required: true
...
    PaginationUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/User"
        totalPages:
          type: integer
          format: int32
          example: 50
        totalElements:
          type: integer
          format: int64
          example: 9266
        size:
          type: integer
          format: int32
          example: 100
        number:
          type: integer
          format: int32
          example: 0
...
    Order:
      type: object
      required:
        - quantity
        - creationDate
        - id_item
        - id_user
      properties:
        id:
          type: integer
        quantity:
          type: integer
        creationDate:
          type: string
          format: date-time
        id_item:
          type: integer
          format: int64
        id_user:
          type: integer
          format: int64
...
    PaginationOrderResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        totalPages:
          type: integer
          format: int32
          example: 50
        totalElements:
          type: integer
          format: int64
          example: 9266
        size:
          type: integer
          format: int32
          example: 100
        number:
          type: integer
          format: int32
          example: 0
...
    StockMovement:
      type: object
      required:
        - quantity
        - creationDate
        - id_item
      properties:
        id:
          type: integer
          example: 149467
          required: true
        quantity:
          type: integer
        id_item:
          type: integer
          format: int64
        creationDate:
          type: string
          format: date-time
...
    PaginationStockMovementResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/StockMovement"
        totalPages:
          type: integer
          format: int32
          example: 50
        totalElements:
          type: integer
          format: int64
          example: 9266
        size:
          type: integer
          format: int32
          example: 100
        number:
          type: integer
          format: int32
          example: 0